# build(deps): update typescript to 3.5 RC