--- a/package.json
+++ b/package.json
@@ -73,7 +73,7 @@
     "power-assert": "^1.6.1",
     "prettier": "^1.17.1",
     "rimraf": "^2.6.3",
-    "typescript": "^3.4.5"
+    "typescript": "^3.5.0-rc"
   },
   "homepage": "https://github.com/teppeis/duck",
   "repository": {