> [Announcing TypeScript 3.5 RC | TypeScript](https://devblogs.microsoft.com/typescript/announcing-typescript-3-5-rc/)

Build performance improvement of tsc (v3.4 → v3.5 RC)
- clean build: 3.3s → 3.0s
- incremental build: 1.7s → 1.4s

Also fixed the typing of `faast()` for new tsc that detects type errors more accurately.